Fix. Area -> Long type

+++ backend/antenna/src/main/java/com/ssafy/antenna/domain/antenna/dto/PostAntennaReq.java
@@ -5,7 +5,7 @@ import jakarta.validation.constraints.Size;
 
 public record PostAntennaReq(
         @Max(value =20, message = "AREA_SIZE_ERROR")
-        int area,
+        Long area,
         double lng,
         double lat) {
 }
